// src/utils/enum/wallet/wallet-type.enum.ts
export enum WALLET_TYPE {
  DEPOSIT = 'DEPOSIT',
  WITHDRAW = 'WITHDRAW',
  REFUND = 'REFUND',
  DRIVER_CREATE_TRIP_FEE = 'DRIVER_CREATE_TRIP_FEE',
  DRIVER_DONE_TRIP_REFUND = 'DRIVER_DONE_TRIP_REFUND',
  DRIVER_DONE_TRIP_FEE = 'DRIVER_DONE_TRIP_FEE',
  PASSENGER_APPROVE_FEE = 'PASSENGER_APPROVE_FEE',
  SYSTEM_GAIN_CREATE_TRIP_FEE = 'SYSTEM_GAIN_CREATE_TRIP_FEE',
  SYSTEM_GAIN_PASSENGER_APPROVE_FEE = 'SYSTEM_GAIN_PASSENGER_APPROVE_FEE',
  SYSTEM_GAIN_DONE_TRIP_FEE = 'SYSTEM_GAIN_DONE_TRIP_FEE',
  DRIVER_BUY_PACKAGE = 'DRIVER_BUY_PACKAGE',
  SYSTEM_GAIN_PACKAGE_FEE = 'SYSTEM_GAIN_PACKAGE_FEE',
  SYSTEM_REFUND_TO_DRIVER_DONE_TRIP = 'SYSTEM_REFUND_TO_DRIVER_DONE_TRIP',
  SYSTEM_REFUND_TO_PASSENGER = 'SYSTEM_REFUND_TO_PASSENGER',
}